Revert "Added Node.js to Docker and linked with existing PostgreSQL Docker; c…"
