--- .idx/integrations.json
@@ -1,3 +1,0 @@
-{
-  "firebase_hosting": {}
-}
