+++ hardhat.config.js
@@ -5,7 +5,7 @@ require("dotenv").config();
 require("@nomiclabs/hardhat-ethers");
 
 const { API_URL, PRIVATE_KEY } = process.env;
-console.log(API_URL);
+
 module.exports = {
   solidity: "0.8.24",
   networks: {
