+++ package.json
@@ -4,27 +4,24 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "hardhat compile"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@nomicfoundation/hardhat-toolbox": "^2.0.1",
     "@openzeppelin/contracts": "^5.0.2",
-    "dotenv": "^16.4.5",
+    "dotenv": "^16.0.3",
     "ethers": "^5.7.1",
     "express": "^4.18.2",
     "express-fileupload": "^1.4.0",
     "fs-extra": "^11.2.0",
     "hardhat": "^2.12.7",
     "html-pdf": "^3.0.1",
-    "locate-chrome": "^0.1.1",
     "moment-timezone": "^0.5.45",
     "path": "^0.12.7",
     "pdf-lib": "^1.17.1",
     "pg": "^8.12.0",
-    "puppeteer": "^22.12.0",
     "qrcode": "^1.5.3"
   }
 }
