+++ server.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 const express = require('express');
-const initializeDatabase = require('./src/resources/initializeDB');
+const initializeDatabase = require('../nodejs-smartcontract/src/resources/initializeDB');
 const app = express();
 const PORT = process.env.PORT || 3000;
 
@@ -10,8 +10,8 @@ const startServer = async () => {
     await initializeDatabase(); // Ensure the table is created
 
     // Import routes
-    const academicRoutes = require("./src/routes/academicRoutes");
-    const factoryRoutes = require("./src/routes/factoryRoutes");
+    const academicRoutes = require("../nodejs-smartcontract/src/routes/academicRoutes");
+    const factoryRoutes = require("../nodejs-smartcontract/src/routes/factoryRoutes");
 
     // Use routes
     app.use('/api/academic', academicRoutes);
