+++ src/controllers/academicController.js
@@ -2,7 +2,6 @@ const pool = require('../resources/db');
 const fs = require('fs-extra');
 const generateCertificatePDF = require('../utils/generateCertificatePDF');
 require("dotenv").config();
-const path = require('path');
 
 const academicAddress = process.env.CONTRACT_ADDRESS_ACADEMIC_ISTER;
 
@@ -35,9 +34,7 @@ exports.issueCertificate = async (req, res) => {
         const pdfBuffer = await generateCertificatePDF(name, documentIdentification, course, description, humanReadableTimestamp, tokenId, receipt.transactionHash, academicAddress);
         
         const pdfFileName = `certificate_${name}-${tokenId}-${timestamp}.pdf`;
-        // const filePath = `./certificatesInPDF/${pdfFileName}`; 
-        // New absolute path
-        const filePath = path.join(__dirname, '..', 'certificatesInPDF', pdfFileName); 
+        const filePath = `./certificatesInPDF/${pdfFileName}`; 
         await fs.writeFile(filePath, pdfBuffer);
     
         responseHandler.success(res, {
