+++ src/resources/db.js
@@ -2,8 +2,7 @@ const { Pool } = require('pg');
 
 const pool = new Pool({
     user: 'ister',
-    host: 'postgres',
-    // host: 'idx-academic-ister-1719265372037.cluster-m7tpz3bmgjgoqrktlvd4ykrc2m.cloudworkstations.dev',
+    host: 'localhost',
     database: 'academic_certificates',
     password: '1ster',
     port: 5432,
