--- src/utils/bak.certificate-template.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            line-height: 1.6;
-            padding: 50px;
-        }
-        .certificate-title {
-            text-align: center;
-            font-size: 24px;
-            font-weight: bold;
-            margin-bottom: 20px;
-        }
-        .logo {
-            display: block;
-            margin: 0 auto;
-            width: 200px; 
-            height: auto; 
-            margin-bottom: 20px;
-        }
-        .details {
-            margin-top: 30px;
-            font-size: 14px;
-        }
-        .qr-code {
-            margin-top: 20px;
-            display: block;
-            text-align: center;
-        }
-    </style>
-    <title>Academic Certificate</title>
-</head>
-<body>
-    <div class="certificate-title">Academic Certificate</div>
-    <img class="logo" src="data:image/jpeg;base64,{{Logo}}" alt="Institution Logo">
-    <div class="details">
-        <p><strong>Name:</strong> {{name}}</p>
-        <p><strong>Document Identification:</strong> {{documentIdentification}}</p>
-        <p><strong>Course:</strong> {{course}}</p>
-        <p><strong>Description:</strong> {{description}}</p>
-        <p><strong>Issued At:</strong> {{issuedAt}}</p>
-        <p><strong>Token ID:</strong> {{tokenId}}</p>
-        <p><strong>Transaction Hash:</strong> {{transactionHash}}</p>
-    </div>
-    <div class="qr-code">
-        <img src="data:image/png;base64,{{transactionHashQRBase64}}" alt="Transaction Hash QR Code">
-    </div>
-</body>
-</html>
