--- src/utils/styles.css
@@ -1,106 +1,0 @@
-/* Estilos para el certificado web */
-/* @media print { */
-* {
-    -webkit-print-color-adjust: exact !important;   /* Chrome, Safari 6 – 15.3, Edge */
-    color-adjust: exact !important;                 /* Firefox 48 – 96 */
-    print-color-adjust: exact !important;           /* Firefox 97+, Safari 15.4+ */
-}
-
-@page {
-    size: A4 landscape;
-    margin: 0;
-}
-    
-html, body {
-    margin: 0;
-    padding: 0;
-    width: 100%;
-    height: 100%;
-    color: #28348A; /* Color de fondo sin comillas */
-    font-family: Arial, sans-serif;
-}
-
-body {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-/* body {
-    margin: 0;
-    padding: 0;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-family: Arial, sans-serif;
-    color: #28348A; 
-} */
-
-.container {
-    width: 100%;
-    height: 100%;
-    padding-bottom: 100vh;
-    /* border: #28348A solid; */
-    min-height: 50vh;
-    background: #D1EBF9;
-    /* background-image: url('./img/background.png');
-    background-size: cover;
-    background-position: top;
-    background-repeat: no-repeat; */
-    border-radius: 10px;
-    padding: 20px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
-    text-align: center;
-    display: flex;
-    justify-content: center;
-}
-
-.name-participant-section {
-    /* background: red; */
-    margin-top: 15rem;
-    margin-left: 25rem;
-    /* border: #28348A solid; */
-    max-width: 50%;
-    padding-bottom: 25rem;
-}
-
-.qr-code {
-    display: flex;
-    justify-content: start;
-    margin-top: 3rem;
-}
-
-.participant-name {
-    font-family: "Playball", cursive !important;
-    font-weight: 400;
-    font-style: normal;
-    color: #28348A;
-    font-size: 55px;
-}
-
-h1 {
-    font-size: 3em;
-    margin-bottom: 20px;
-}
-
-p {
-    font-size: 1.5em;
-}
-
-/* fonts */
-.playball-regular {
-    font-family: "Playball", cursive;
-    font-weight: 400;
-    font-style: normal;
-    color: #28348A; 
-}
-.montserrat-text {
-    font-family: "Montserrat", sans-serif;
-    font-optical-sizing: auto;
-    font-weight: 400;
-    font-style: normal;
-  }
-
-/* } */
